added permissions so that only the author of a thread can see the delete button on their thread. Also protected at the controller level

+++ app/Http/Controllers/ThreadController.php
@@ -117,12 +117,20 @@ class ThreadController extends Controller
    */
   public function destroy(Channel $channel, Thread $thread)
   {
+    if ($thread->id != auth()->id()) {
+      if (request()->wantsJson()) {
+        return response(['status' => 'Permission Denied'], 403);
+      }
+      return redirect('/login');
+    } else {
       $thread->delete();
       if (request()->wantsJson()) {
         return response([], 204);
       }
 
       return redirect('/threads');
+    }
+
   }
 
   public function getThreads(Channel $channel, ThreadFilters $filters)
